Develop

+++ trace_validation_pipeline.py
@@ -4,39 +4,47 @@ import run_impl
 import trace_merger
 import tla_trace_validation
 import argparse
+import ndjson
 
-parser = argparse.ArgumentParser(
-    prog='Trace validation pipeline',
-    description='This program aims to execute a pipeline that validate implementation of KeyValueStore against a formal spec.')
+def read_json(filename):
+    with open(filename) as f:
+        return ndjson.load(f)
 
+def get_files(config):
+    files = ["store.ndjson"]
+    # No files in config
+    return files
+
+parser = argparse.ArgumentParser("")
 parser.add_argument('-c', '--compile', type=bool, action=argparse.BooleanOptionalAction)
-
+parser.add_argument('--config', type=str, required=False, default="conf.ndjson", help="Config file")
 args = parser.parse_args()
 
-print("# Clean up")
+config = read_json(args.config)
+files = get_files(config)
 
-trace_files = [f for f in os.listdir(".") if f.endswith('.ndjson')]
-print(f"Cleanup: {trace_files}")
+# Clean up
+print("# Clean up")
+trace_files = files + ["trace.ndjson"]
+print(f"Cleanup: {files}")
 for trace_file in trace_files:
-    os.remove(trace_file)
+    if os.path.isfile(trace_file):
+        os.remove(trace_file)
 
 if args.compile:
     print("# Compile.\n")
     subprocess.run(["mvn", "package"])
 
+# Run
 print("# Start implementation.\n")
-
 run_impl.run()
 
-print("# Merge trace with config.\n")
+# Merge traces 
+print("# Merge traces.\n")
+trace_merger.run(files, sort=True, remove_meta=True, out="trace.ndjson")
 
-trace_tla = trace_merger.run(["."], config="kvs.ndjson.conf", sort=True)
-# Write to file
-with open("trace-tla.ndjson", "w") as f:
-    f.write(trace_tla)
-
+# Validate trace
 print("# Start TLA+ trace spec.\n")
+tla_trace_validation.run_tla("spec/KeyValueStoreTrace.tla","trace.ndjson",args.config)
 
-tla_trace_validation.run_tla("spec/KeyValueStoreTrace.tla","trace-tla.ndjson")
-
-print("End pipeline.")
+# print("End pipeline.")
